mstar: fb_notifier: Move TP on/off operations to a workqueue
The driver is currently performing operations that take several hundreds
of milliseconds on the blocking mdss notifier chain. This causes the
screen-on time to be too percetible to the user.

This patch adds a workqueue that processes the TP on/off events. The
work is triggered by the mdss notifier chain but processed in a
different one.

Change-Id: I34197492347e31817e078a563e17f85e1c6ddb8e